Update gitconfig
Fix tabs (change to spaces). Set crlf options. Define color options. Add
github options.

The github token should not be made public, so I wrote a quick python
script to access the gnome-keyring to grab the token from there. The
Makefile uses this script to substitute the token into a new gitconfig
file. It does not handle the case where the key does not exist in the
keyring. We also make sure this file is not saved by git.