Remove dir_colors
This commit removes the dir_colors file. First, it was misspelled. Some
systems (default Ubuntu, at least) will `eval` a `.dircolors` file
automatically. Furthermore, dircolor files can come from a variety of
sources, and it may or may not make sense to manage them here. For
example, my current favorite comes from the Solarized theme,
specifically the `seebi/dircolors-solarized` repository on Github. It
makes more sense to me to manage that repository separately from this
one.

If there is a way to sync a single file with another remote, I may do
that instead, in the future.