Add long line highlighting in vim
This commit adds a function to vim to highlight the ends of long lines.
Any column after the 76 gets highlighted in the `ErrorMsg` color but to
serve as a warning, anything after the 73 column gets highlighted in the
`Search` color. This function is enabled by default in (almost) all
cases, but can be toggled with `<F3>`.