Use the solarized color theme for Vim
Before I was setting my terminal emulator to use the solarized theme and
then hoping that Vim would use the right colors. That was a faulty
thinking. This commit sets vim to use the solarized color theme.

Also included is better whitespace detection that should look okay with
or without solarized.