Drop lead . (period) from Xresources #include
When (load/merge)ing Xresources, xrdb or the C preprocessor looks for
files to #include relative to the actual file and not to the symlink in
the home directory. Thus we need to drop the leading . (period) from the
directive to make it load the proper file.

This is either a bug or feature, I can't decide, since it means we don't
have to link Xresources.d in the home directory.