Switch to using urxvtd daemon
We now start a urxvt daemon when X starts set the default terminal to
connect to the running daemon.

If the daemon crashes, all connected clients go down with it and new
clients will fail to connect. There should probably be a wrapper that
will connect to an already running process or start a new daemon and
connect to it.