Move and update xmonad
You can include external custom libraries in XMonad by adding them to
the lib directory relative to xmonad.hs. We do this with the Solarized
library which defines strings to represent the Solarized color theme,
which we make use of in our XMonad configuration.

We also add and the Accordion and tabbed layouts and remove the Full
layout.