Setup bash and vim for ruby/rails development
Besides setting some reasonable spacing options in vim for ruby and
common rails related files, this commit also includes bash-completion
scripts for rails and rake.

Also included is a simple little function for starting (or reconnecting
to) a preconfigured tmux session for rails development. It creates
several windows and starts a rails server, guard, and vim at the rails
root.