Shorten home directory to ~ in rxvt and tmux
My prompt command function uses escape sequences to change the
window/pane title in rxvt/tmux. I had intended for home directory to be
shortened to a ~, but my bash-fu was not strong enough. The tilde in the
replacement string was being expanded, meaning that the home directory
with the home directory. Escaping the ~ fixes that.