Allow xmobarrc to be host specific
Generally, I've been keeping host specific configuration changes on a
branch local to each host. This isn't always that great as the recent
change to my xmobarrc has made apparent; any changes I had to my
xmobarrc on a local branch are now useless. In the future, I am going to
try to avoid have changes on a local branch.