Improve xmobarrc generation
When generating xmobarrc, I was dropping any configurations that matched
the default options. This is completely unnecessary, and complicated the
code. I also removed the extra spaces where appropriate, switched to
using the infix `<>` operator instead of append, used wrap/surround in
my text functions, and cleaned up the imports.