Switch to Solarized light
I've decided to try to switch to the light version of Solarized. I
enjoyed the dark version, but I want to give the light one a chance
again. I was also kind of wanting a change in colors, but couldn't find
anything better than Solarized. Furthermore, most websites have white or
light colored themes, and this way the switch will be less jarring and
my eyes will have much less work to do (which is part of the whole thing
with Solarized). Also in low-light environments, a light theme should
let me lower the brightness on my monitor more giving a modest power
saving bonus.